projects/stock-analyser/requirements.txt
requests>=2.25.0
pandas>=1.3.0
ratelimit>=2.2.1
pathlib>=1.0.1
python-dotenv>=1.0.0